--- a/include/spl/sys/mutex.h
+++ b/include/spl/sys/mutex.h
@@ -179,7 +179,4 @@ spl_mutex_lockdep_on_maybe(kmutex_t *mp)			\
 	/* NOTE: do not dereference mp after this point */	\
 }
 
-int spl_mutex_init(void);
-void spl_mutex_fini(void);
-
 #endif /* _SPL_MUTEX_H */