### Motivation and Context

Resolve Linux 5.3 build failures.  @kusumi would you mind reviewing this.

### Description

This PR contains three changes but only fa0047f0cc50a8dd3f8da0f82a88aefd111200d4 is required to resolve the build issue with 5.3.  The other two changes were included as additional cleanup in order to avoid potential future problems.  Over the years the rwsems have seen numerous changes and going forward it's desirable to be more resilient to kernel changes.

* fa0047f0cc50a8dd3f8da0f82a88aefd111200d4 Linux 5.3 compat: rw_semaphore owner
* e4d18399c30b57cb65aac37f4cd36c6d5e1de400 Linux 5.3 compat: retire rw_tryupgrade()
* 4459f2b44812f58f9e4da5f3a276ed8e8a16e612 Retire unused spl_{mutex,rwlock}_{init_fini}

### How Has This Been Tested?

Locally ran portions of the ZTS using the 5.3.0-0.rc0.git4.1.vanilla.knurd.1.fc30.x86_64 kernel.  Pending CI results to verify other kernels.  No significant performance regression was observed after disabling `rw_tryupgrade`.

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Performance enhancement (non-breaking change which improves efficiency)
- [x] Code cleanup (non-breaking change which makes code smaller or more readable)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation (a change to man pages or other documentation)

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the ZFS on Linux [code style requirements](https://github.com/zfsonlinux/zfs/blob/master/.github/CONTRIBUTING.md#coding-conventions).
- [x] I have updated the documentation accordingly.
- [x] I have read the [**contributing** document](https://github.com/zfsonlinux/zfs/blob/master/.github/CONTRIBUTING.md).
- [ ] I have added [tests](https://github.com/zfsonlinux/zfs/tree/master/tests) to cover my changes.
- [x] All new and existing tests passed.
- [x] All commit messages are properly formatted and contain [`Signed-off-by`](https://github.com/zfsonlinux/zfs/blob/master/.github/CONTRIBUTING.md#signed-off-by).