@ahrens 2019-07-12
<3 the code simplification here, but have you evaluated the potential performance implications of never upgrading rwlocks?