@behlendorf 2019-07-13
> have you evaluated the potential performance implications of never upgrading rwlocks?

Yes, via code inspection of the callers.  There are 4 callers of `rw_tryupgrade()`, of which I believe only one, `dmu_zfetch()`, may have an impact on performance.  They are:

1) `dmu_zfetch()` - When creating a new zstream in `dmu_zfetch()`, if the code is unable to upgrade the `zf->zf_rwlock` the dnode's `dn_struct_rwlock` is dropped and reacquired, then the code rescans for a matching prefetch stream.  This behavior could be improved.

2) `zap_tryupgradedir()` - Most callers already hold the lock as a writer, and zap upgrades aren't expected to be a high frequency events so any impact here is expected to be minimal.

3) `dnode_new_blkid()` - Negligible (if any) change in behavior, on upgrade failure the read lock is dropped and immediately acquired as a writer.

4) `zfs_fuid_find_by_domain()` - Results in one additional call to `avl_find()` on failure, only reachable from `zfs_set_userquota()`.

If you can suggest a test case to stress `dmu_zfetch()` I'm happy to perform the testing.  Though, I'd like to address any loss in performance by updating the caller and not re-introducing `rw_tryupgrade()`.