@ahrens 2019-07-15
@behlendorf A good workload to exercise this is random reads or writes of 8k blocks (on recordsize=8k) with several threads (>=8 threads on an 8-CPU system worked well for me).

I happen to be working on some changes to the dmu_zfetch locking anyway (because even the current rwlock implementation has extreme performance pathologies).  So it's probably OK if this regresses performance of that somewhat.