@behlendorf 2019-07-17
@ahrens using the suggested workload (random 8k reads, 8k recordsize) I compared performance with and without this PR.  Tested using an ec2 c5d.xlarge instance type (4 vcpu, local NVMe instance storage, Amazon Linux 2,  4.14.128-112.105.amzn2.x86_64 kernel).  Based on these quick results I don't see any appreciable impact on performance.

processes |master (MiB/s) | patched (MiB/s)
--- | --- | --- 
1 | 52.9 | 52.3 |
2 | 109.0 | 109.5 |
4 | 189.9 | 191.1 |
8 | 188.3 | 186.7 |
16 | 186.9 | 187.4 |
32 | 189.2 | 191.9 |
64 | 190.3 | 188.4 |
128 | 189.8 | 188.6 |
256 | 191.0 | 190.1 |
512 | 195.2 | 192.8 |


<details open>
<summary>8k-read.fio</summary>
<pre>
[global]
filename=file
group_reporting=1
fallocate=0
overwrite=0
rw=randread
directory=/tank/fs/
runtime=10
bs=8192
ioengine=libaio
iodepth=1
sync=0
direct=1
size=32G
numjobs=${JOBS}

[job]
</pre>
</details>

Note: I also didn't observe any pathological degradation of perform when threads > ncpus.